requirements.txt
bark_ssg==1.3.4
decord==0.6.0
diffusers==0.15.0
einops==0.7.0
imageio==2.28.0
ipython==8.14.0
librosa==0.10.1
mmcv==2.1.0
moviepy==1.0.3
natsort==8.3.1
nltk==3.8.1
numpy==1.23.5
omegaconf==2.3.0
openai==0.27.8
opencv_python==4.7.0.72
Pillow==9.4.0
Pillow==10.2.0
pytorch_lightning==2.0.2
rotary_embedding_torch==0.2.3
soundfile==0.12.1
torch==2.0.0
torchvision==0.15.0
tqdm==4.65.0
transformers==4.28.1
xformers==0.0.19
